% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{add_labels}
\alias{add_labels}
\title{Read labels from a varlist sheet and add them to the given data table.}
\usage{
add_labels(table_data, varlist)
}
\arguments{
\item{table_data}{A tibble of IPEDS data.}

\item{varlist}{A tibble of an IPEDS metadata file's `varlist` sheet.}
}
\value{
The modified `table_data`.
}
\description{
Read labels from a varlist sheet and add them to the given data table.
}
