% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_metadata}
\alias{download_metadata}
\title{Download the metadata for a given IPEDS table.}
\usage{
download_metadata(table_id)
}
\arguments{
\item{table_id}{A table ID.}
}
\value{
The path to the downloaded zip file of that table's metadata.
}
\description{
Download the metadata for a given IPEDS table.
}
