% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_codebook}
\alias{extract_codebook}
\title{Load the codebook from an IPEDS zip file.}
\usage{
extract_codebook(local_path)
}
\arguments{
\item{local_path}{The path to a downloaded zip file.}
}
\value{
A tibble of data.
}
\description{
Load the codebook from an IPEDS zip file.
}
