% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_csv}
\alias{extract_csv}
\title{Load the data CSV from an IPEDS zip file.}
\usage{
extract_csv(local_path, column_types)
}
\arguments{
\item{local_path}{The path to a downloaded zip file.}

\item{column_types}{A column specification as returned by `readr::cols`.}
}
\value{
A tibble of data.
}
\description{
Load the data CSV from an IPEDS zip file.
}
