% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_varlist}
\alias{extract_varlist}
\title{Load the column type/title metadata from an IPEDS zip file.}
\usage{
extract_varlist(local_path)
}
\arguments{
\item{local_path}{The path to a downloaded zip file.}
}
\value{
A tibble of data.
}
\description{
Load the column type/title metadata from an IPEDS zip file.
}
