% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_column_types}
\alias{get_column_types}
\title{Build a full column specification from an IPEDS varlist.}
\usage{
get_column_types(varlist)
}
\arguments{
\item{varlist}{A tibble of an IPEDS metadata file's `varlist` sheet.}
}
\value{
A column specification as returned by `readr::cols`.
}
\description{
Build a full column specification from an IPEDS varlist.
}
