% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_table.R
\name{ipeds_table}
\alias{ipeds_table}
\title{Load a table's data and important metadata directly from the IPEDS website.}
\usage{
ipeds_table(table_id)
}
\arguments{
\item{table_id}{A table ID.}
}
\value{
A list containing three tibbles: $table, $variables and $codebook.
}
\description{
Load a table's data and important metadata directly from the IPEDS website.
}
