% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_metadata}
\alias{download_metadata}
\title{Download the metadata for a given IPEDS table.}
\usage{
download_metadata(table_id, local_dir)
}
\arguments{
\item{table_id}{A table ID.}

\item{local_dir}{A path to a directory where the raw data file should be
saved.}
}
\value{
The path to the downloaded zip file of that table's metadata.
}
\description{
Download the metadata for a given IPEDS table.
}
