% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds_table.R
\name{ipeds_table}
\alias{ipeds_table}
\title{Load the data CSV from an IPEDS zip file.}
\usage{
ipeds_table(table_id)
}
\arguments{
\item{table_id}{A table ID.}
}
\value{
A list containing two tibbles: $table and $codebook.
}
\description{
Load the data CSV from an IPEDS zip file.
}
